# spec/open_api_spec.yml
swagger: "2.0"
info:
  description: "A sample Scala project using Spring Boot and providing examples of validation and end-point testing.
  for more information on the OpenAPI Specification (aka Swagger) used herein see
  <a href=http://openapi-specification-visual-documentation.apihandyman.io/>OpenAPI Visual Documentation</a>"
  version: 1.0.0
  title: Spring Boot Scala REST API Application Example
  contact:
    email: "Joseph.Simone@nexteraenergy.com"
basePath: /
...
schemes:
  - http
paths:
...
  /hello0:
    get:
      tags:
        - Hello World
      produces:
        - text/plain
      parameters:
        - name: name
          in: query
          description: Name to say hello to
          required: false
          type: string
      responses:
        "200":
          description: successful operation
...
  /hello1.1:
    get:
      tags:
        - Hello World
      produces:
        - text/plain
      parameters:
        - name: name
          in: query
          description: Name to say hello to
          required: true
          type: string
          default: World
      responses:
        "200":
          description: successful operation
  /hello1.2:
    get:
      tags:
        - Hello World
      produces:
        - text/plain
      parameters:
        - name: num
          in: query
          description: Number to say hello to
          required: true
          type: integer
      responses:
        "200":
          description: successful operation
...
  /hello2.1/{num}:
    get:
      tags:
        - Hello World
      produces:
        - text/plain
      parameters:
        - name: num
          in: path
          description: Number to say hello to
          required: true
          type: integer
      responses:
        "200":
          description: successful operation
...
  /hello5:
    get:
      tags:
        - Hello World
      description: An attempt at validating the query parameters via @NotEmpty @Size(min=2, max=30) @RequestParam and @Valid annotations
      produces:
        - text/plain
      parameters:
        - in: query
          name: name
          description: Person name to say hello to
          required: false
          type: string
      responses:
        "200":
          description: successful operation
...